backend_db/requirements.txt
Django==4.2.27
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==10.4.0
daphne>=4.0,<5.0